PUSHBOT RED
Added pushbot red and removed the unnecasary meep meep
